chore(server): file tree
