+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration-v2/utils/flatten-field-metadata-comparator.util.ts
@@ -4,7 +4,7 @@ import { isDefined } from 'twenty-shared/utils';
 
 import { FlattenFieldMetadata } from 'src/engine/workspace-manager/workspace-migration-v2/types/flatten-field-metadata';
 import { FromTo } from 'src/engine/workspace-manager/workspace-migration-v2/types/from-to.type';
-import { UpdateFieldAction } from 'src/engine/workspace-manager/workspace-migration-v2/types/workspace-migration-field-action-v2';
+import { UpdateFieldAction } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/types/workspace-migration-field-action-v2';
 import { transformMetadataForComparison } from 'src/engine/workspace-manager/workspace-sync-metadata/comparators/utils/transform-metadata-for-comparison.util';
 
 const workspaceMigrationFieldInputPropertiesToCompare = [
