+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration-v2/utils/flatten-index-metadata-comparator.util.ts
