+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration-v2/utils/flatten-object-metadata-comparator.util.ts
@@ -4,7 +4,7 @@ import { assertUnreachable } from 'twenty-shared/utils';
 
 import { FlattenObjectMetadata } from 'src/engine/workspace-manager/workspace-migration-v2/types/flatten-object-metadata';
 import { FromTo } from 'src/engine/workspace-manager/workspace-migration-v2/types/from-to.type';
-import { UpdateObjectAction } from 'src/engine/workspace-manager/workspace-migration-v2/types/workspace-migration-object-action-v2';
+import { UpdateObjectAction } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/types/workspace-migration-object-action-v2';
 import { transformMetadataForComparison } from 'src/engine/workspace-manager/workspace-sync-metadata/comparators/utils/transform-metadata-for-comparison.util';
 
 const workspaceMigrationObjectInputPropertiesToCompare = [
