+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/utils/get-workspace-migration-v2-field-actions.ts
@@ -2,7 +2,7 @@ import {
   CreateFieldAction,
   DeleteFieldAction,
   FieldAndObjectMetadataWorkspaceMigrationInput,
-} from 'src/engine/workspace-manager/workspace-migration-v2/types/workspace-migration-field-action-v2';
+} from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/types/workspace-migration-field-action-v2';
 
 export const getWorkspaceMigrationV2FieldCreateAction = ({
   flattenFieldMetadata,
