+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/workspace-migration-v2-index-actions-builder.ts
@@ -1,6 +1,6 @@
-import { WorkspaceMigrationIndexActionV2 } from 'src/engine/workspace-manager/workspace-migration-v2/types/workspace-migration-index-action-v2';
+import { compareTwoFlattenedIndexMetadata } from 'src/engine/workspace-manager/workspace-migration-v2/utils/flatten-index-metadata-comparator.util';
+import { WorkspaceMigrationIndexActionV2 } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/types/workspace-migration-index-action-v2';
 import { UpdatedObjectMetadataDeletedCreatedUpdatedIndexMatrix } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/utils/compute-updated-object-metadata-deleted-created-updated-index-matrix.util';
-import { compareTwoFlattenedIndexMetadata } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/utils/flatten-index-metadata-comparator.util';
 import {
   getWorkspaceMigrationV2CreateIndexAction,
   getWorkspaceMigrationV2DeleteIndexAction,
